ide: Automatic setup refactored to use registry
git-svn-id: svn://ultimatepp.org/upp/trunk@10625 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/Vcont.hpp
@@ -725,7 +725,7 @@ BiVector<T>::BiVector(std::initializer_list<T> init)
 	start = items = alloc = 0; vector = NULL;
 
 	start = 0;
-	alloc = items = init.size();
+	alloc = items = (int)init.size();
 	if(!alloc) {
 		vector = NULL;
 		return;
